requirements.txt
dvc==3.0.0
mlflow==2.0.0
scikit-learn==1.3.0
pandas==2.0.0
numpy==1.24.0
pyyaml==6.0
joblib==1.3.0